Correct unchecking option checkboxes
If a checkbox is not checked, the value is not present in the form 
parameters sent from the client. In this case, when we sanitize, the 
option would be restored by the values in the old array.

The fix consists in not including the old parameter in the newly 
sanitize array if the corresponding key is not present in the input 
array.